requirements.txt
numpy
pandas
scikit-learn
xgboost
lightgbm
torch
torchvision
torchaudio
pyyaml
matplotlib
scipy
tqdm
jupyter
notebook
ruff
mypy
pytest
shap
pydantic
